[HACK][WIP][DO NOT MERGE] 16k: Move code from writeable pages.
The original hack to load 4KB aligned shared libraries relies on
mapping everything as RWX. This is a significant security issue.

This work-in-progress solution aims to remove the RWX regions. When
executable code and RW data ends up on the same 16KB page, the
executable parts are copied to a new page. The old page is made
non-executable. A custom signal handler is installed. When the program
attempts to execute code from the old page, a SIGSEGV is raised. The
signal handler detects that it came from the old page, and redirects
execution to the new page.

On the new page, the PC relative offsets have to be updated. Relative
branches that point into the same page can be kept as is. Everything
else has to be updated, so they point to the old page. If the adjusted
offset fits into the instruction, the instruction is patched.
Otherwise, a trap is inserted. A signal handler is installed to handle
these traps. When it detects that a signal comes from one of the
trapped instructions, it adjusts registers as appropriate. ADR/P
instructions are always patched or trapped, to make sure they point to
the old page. This way everything in the new page can be patched
indiscriminately, as all the data will be read from the old page.

Note that this means that the signal handlers installed by the dynamic
linker should always run first. The way to do this properly is
work-in-progress.

Co-authored-by: Daniel Kiss <daniel.kiss@arm.com>
Co-authored-by: Gabor Kertesz <gabor.kertesz@arm.com>
Co-authored-by: Tamas Zsoldos <tamas.zsoldos@arm.com>

Change-Id: I54d792a8f279d0af5950ba27d9ff3104ad0e69ea