Revert "Modify argument parser for multiple arguments."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0f202070-b0d8-4cc2-9724-fd2302334183). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia5436da637c6aaec2e0d1985b40576be01cf70ea