Revert "Modify argument parser for multiple arguments."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5473ba6f-3fa0-465c-88b7-c28313dba817). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9444852efdcea627bb2e478193db19a46e7907b8