Revert "Remove __RENAME_STAT64."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eda2c009-9a24-4d6d-9d78-38ebb7c9a05e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie95dd63c8ef62cee5c3d0b682fff06206172e9aa