fix BionicSmallObjectAllocator race condition while do alloc/free
Once the program is executed, the linker will link a large number of
libraries and search for symbols.
Some libraries will start threads after initialization.
In this situation, strings such as symbol name and log will utilize the
bionic allocator.
The bionic allocator will allocate the corresponding
BionicSmallObjectAllocator based on the length of the string.
Due to a large number of strings with lengths ranging from 32 to 64
bytes,
it is more prone to race conditions in the 64-byte block size allocator.
This can result in allocating already used memory blocks to users,
leading to subsequent program crashes.

Bug: 290318196
Test: Reboot test pass
Change-Id: Idbdc9b83ec2d4f00be010637bd269312ae74e0b1