getentropy should be in <unistd.h>, not <sys/random.h>
Per https://man7.org/linux/man-pages/man3/getentropy.3.html, the man
page bionic itself cites in its comments, getentropy should be provided
by <unistd.h>. This also matches what the BSDs do:

https://man.freebsd.org/cgi/man.cgi?query=getentropy&sektion=3&format=html
https://man.openbsd.org/getentropy.2

As well as musl:
https://git.musl-libc.org/cgit/musl/tree/include/unistd.h?id=25e6fee27f4a293728dd15b659170e7b9c7db9bc#n183

POSIX will likely place it there too:
https://www.austingroupbugs.net/view.php?id=1134

macOS and Fuchsia place it in <sys/random.h>, but they are outliers.
Using a standard location will improve portability and avoid needing
workarounds in BoringSSL.

This CL moves it to a header in bits/ so that sys/random.h and unistd.h
can both expose it. unistd.h is the standard location and sys/random.h
is for compatibility.

BYPASS_INCLUSIVE_LANGUAGE_REASON=Above URLs are not hosted by Android and reference the name of a command-line utility, short for 'manual', as in instruction manual

Test: treehugger
Change-Id: Id2d6b6ea09d814e5ba2cb117a7af2c74861148fb