Remove __BIONIC_NO_PAGE_SIZE_MACRO
Just don't define PAGE_SIZE when the product configures it.
This cl changes how the configuration is done, now it's done via the
"page_size_agnostic" soong config variable in the "bionic" namespace.

page.h will be generated to either include the PAGE_SIZE definition or
not. This allows us to not have a separate __BIONIC_NO_PAGE_SIZE_MACRO
macro.

Bug: 295910468
Test: m
Change-Id: I435a1b3477d22bc01e62c6138b94d934bbf0adb6