CFIShadow: Convert class to namespace
libdl does cannot use the STL so it does not have access to
symbol __cxa_guard_acquire/__cxa_guard_release needed for
thread safe statics.

All member of CFIShadow class are static, convert this to
namespace instead to avoid the issue described above.

This is a preparatory change for the next patch in this series.

Bug: 296275298
Test: m
Change-Id: I5e8d8393c7430516dfa18035c227832d7fc67890
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>