CFIShadow: Convert class to namespace
libdl does cannot use the STL so it does not have access to
symbol __cxa_guard_acquire/__cxa_guard_release needed for
thread safe statics.

All member of CFIShadow class are static, convert this to
namespace instead to avoid the issue described above.

This is a preparatory change for the next patch in this series.

Bug: 296275298
Test: m
Chnage-Id: I577d2fab614df7b9386cf72fc58d3f1feef0267d
Change-Id: Ib7e2dbb7140868bbf9c31b0063016023fab7f926
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>