Revert "bionic: Fix typo in the argument passed to getauxval()"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/46228b70-2417-4cd8-963a-d9426a0703a5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7f8c6950348e29bd9d2aed644b0d57c6df63c9c3