Add a GWP-ASan tests that unpoisons the device
If you ctrl+C the bionic-unit-tests at juuuust the wrong time, when a
GWP-ASan test is running that uses system properties, you can end up not
unsetting them and poisioning the device.

Using the assumption that someone's running bionic-unit-tests multiple
times, and if they ctrl+C they're going to rerun it again, we can setup
a test that looks for a magic value and unpoisons the device.

I'm not sure we can effectively handle someone that ctrl+C's the test
and then doesn't run bionic-unit-tests again. I'm very wary of
installing a SIGINT handler as it may interact poorly with signal-based
tests in bionic.

Bug: 303122139
Test: Add a sleep(10) in one of the poisining tests, ctrl+c, then re-run
bionic-unit-tests and make sure it gets unpoisoned.

Change-Id: I2a22e7231dea63939f316e65173ddddfa509d117