Add android_{get,set}_exit_mode.
Calling exit in a multi-threaded process is a recipe for creating a
dozen different tombstones with different stacks due to destruction of
globals, and it's a pain in the ass to dig through the tombstone's other
thread stacks to find a thread blocked in exit to figure out the root
cause.

Add some flags to allow configuration of exit's behavior, including
having it act as _exit, generate a tombstone, or immediately abort.

Test: bionic_unit_tests on sdk_phone_x86_64-userdebug
Change-Id: I934cb68f39199fdb1730d563adf8bc34c935d3d4