Add android_mallopt M_GET_DECAY_TIME_NON_ZERO.
The bionic benchmarks set the decay time in various ways, but
don't necessarily restore it properly. Add a new method for
getting the current decay time and then a way to restore it.

Right now the assumption is that the decay time is set to zero,
but in the near future that assumption might be different. Therefore
using this method will future proof the code.

Bug: 302212507

Test: Unit tests pass.
Test: Ran bionic benchmarks that were modified.
Change-Id: Ia77ff9ffee3081c5c1c02cb4309880f33b284e82