Support enabling the allocator decay timer.
In order to avoid performance differences between shell spawned
command-line tools and native code running in an app, create
a method for all command-line tools run from the shell to automatically
enable the decay timer.

This should be temporary while Scudo works on a method to change
the default to always enable the decay time but not incur a permanent
RSS increase.

Bug: 302212507

Test: Set the variable to 1 and verified the decay time is enabled.
Test: Set the variable to "" and also unset it and verified the decay
Test: time is not enabled.
Change-Id: I22a47f70f1ce1205c16195532c54b2bf9403e9cd