Avoid calls to free() before allocator init.
Calling free() during system property init results in premature
allocator initialization.

One of the consequences of that is, on MTE devices, mapping of some of
the scudo primary regions as PROT_MTE regardless of the process intent.
This has a performance impact (extra memory traffic), but does not
affect functional correctness.

The test verifies that there are no stray PROT_MTE mappings when MTE is
not enabled in the binary.

Test: adb shell setprop arm64.memtag.bootctl memtag
      adb reboot
      bionic-unit-tests --gtest_filter=MemtagNoteTest.SEGV/*
Bug: 309698651
Change-Id: I6c7733d8799537d898c97b00d494ce6591cf44d9