Revert "generate_notice.py: don't check for strings we don't use..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/988ef4d8-0218-48cf-a0c8-12e4db797422). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I279f95affe2b3e1f374fa35cb915ca3fd3576e2f