Revert "generate_notice.py: don't check for strings we don't use..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/73fbd582-3a9e-4161-92b3-9d18070b2d28). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7e16f3b64c737977071c71300fb5803a63de6861