Use libz_static in static_libs of linker
This is a workaround to avoid making libz available to
com.android.runtime. If we make libz available to runtime apex, coverage
builds run into linking issues for executables that depend on libz.
```
ld.lld: error: undefined reference due to --no-allow-shlib-undefined: __umoddi3@LIBC_PRIVATE
>>> referenced by out/soong/.intermediates/external/zlib/libz/android_x86_silvermont_shared_cov/libz.so
```

Soong uses equality of apex_available for stub vs implementation
selection. In the above example, the executable is linking against a
stubbed version of libc, but against a version of libz that was linked
against an implementation version of libc

Test: presbumits
Bug: 281077552
Change-Id: Ia2f1fd3140d03733504a7b78161d75f54469d691