<ctype.h>: more optimization.
clang is pretty good at turning something like

    (c >= '0' && c <= '9')

into the equivalent of an array bounds check:

    ((unsigned c)-'0' < 10).

Unfortunately, it struggles if you combine two or more of these
comparisons.

So although it generates [afaik] optimal code for isdigit(), and can
separately do the same for

    (c >= 'a' && c <= 'f'),

or

    (c >= 'A' && c <= 'F'),

or even

    (_tolower(c) >= 'a' && _tolower(c) <= 'f'),

when you combine them (as in isxdigit()) it no longer recognizes any of
these as something it can optimize and puts all the branches back in.

To avoid this, let's add a macro that does the optimized range check and
use that throughout <ctype.h>.

This has a tiny impact in our arm64 benchmarks, which were already at the
edge of our ability to measure (basically the isalpha() and iscntrl()
failure cases are now the same "as fast as our clock resolution can
measure", whereas before they were slightly slower than all the other
<ctype.h> benchmarks).

For arm32 (which is worse across the board, even on the same cpu),
the impact is a bit larger --- we go from 2.22ns for every slow case
and 1.66ns for every fast case to 2.05ns for the slow cases and 1.54ns
for the fast cases. (The lack of any conditional instructions in
arm32 means that all the routines are about twice the number of
instructions.)

Only ispunct() still has an explicit branch for arm64; all the arm32
functions lose their branches, if you don't count the `bxlo lr` in
the same part of ispunct().

Test: treehugger
Change-Id: I663ff7e31b0f4173d6dd27f16bb2d8d3485c8e06