Revert "bionic: Introduce ElfReader::ReadPadSegmentNote()"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/93c8df0e-6ae8-4ffe-b2de-1ffb64f7bc95). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1ec0af54ad011abcb964cfd79eaa26c4cf045c70