Revert "use /system/bin/sh for new vendor partition"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ef0ed6dc-6c99-4275-8510-0304fbdd06f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib360ad4a01742d94d5a3592ba7590898733688df