Define __INTRODUCED_IN_LLNDK for LLNDK symbol versioning
Symbols with __INTRODUCED_IN_LLNDK are available as LLNDK only if the
vendor api level set by the RELEASE_BOARD_API_LEVEL is equal to or
newer than the target vendor api level in __INTRODUCED_IN_LLNDK.
This works only for the vendor variants. It does not change symbols
for NDK or core variant libs.

Bug: 302113279
Test: build
Change-Id: If2076605e2db59f50fc943f6e7eb7379bfbb1229