Revert "RELAND: bionic: loader: Extend LOAD segment VMAs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0125a71e-96a0-474b-a96c-7af01d0cc281). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ee7fef8361131f4fec4ddd42f2b5eca17ce00db