bionic: loader: Page out zeroed pages
When the runtime page size < ELF's LOAD segment p_align, the loader
may extend LOAD segment mappings to be contiguous in the virtual address
space. This is done in order to reduce the use of unreclaimable kernel
slab memory for the otherwise necessary gap VMAs.

The zeroing of the extended portion of the mapping may result in
faulting COWed zero pages (anon memory), although this won't be used
again.

While such pages will eventually be reclaimed as the system encounters
regular memory pressure; expedite the process by hinting this to the
kernel with madvise().

Test: bootup/idle/system-processes-memory-direct
Bug: 327600007
Change-Id: I9d013da6d72dbb42fadada028b8922e0f6c813da
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>