Avoid calling mutate unless necessary.
When heapprofd is enabled, dlopen is called to load the heapprofd shared
library, which triggers a call to __pthread_internal_remap_stack_with_mte.
When MTE is enabled, the libc globals are made writable using the mutate
function. However, at the same time the heapprofd code is also trying
to modify the libc globals. Unfortunately, the heapprofd path
takes a while to finish, the remap mutate finishes and makes the
globals read-only. Now the heapprofd tries to modify the globals in
a read-only segment and a crash occurs.

Fix this by only calling mutate if necessary from the remap function.

However, a more thorough fix might be needed to detect this case and
avoid making the globals read-only while they are still being used.

Bug: 328256432

Test: Ran bionic unit tests on a full mte build.
Test: Ran debuggerd unit tests on a full mte build.
Change-Id: I79fca96230c83e87b331adc6a3b547ffec871786