Revert "bionic: loader: Only zero the last partial page in RW se..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/88faae42-6f15-470c-acba-5b2f6f89934a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d32b7faee95dac40630c8cc3614d5ae3433424d