Hide overaligned global address from the compiler.
The address of contents is only guaranteed to be aligned to 4KB on
4KB page size systems, but the compiler was generating code that
assumed it to be aligned to 64KB, which broke on a 4KB page size
system. This probably ought to be fixed in the compiler so it can't
generate code assuming such large alignments (it's hard to see what
useful optimizations are possible by assuming such large alignments
anyway) but for now let's hide the address behind an asm statement
that the compiler can't see through.

Bug: 332534664
Change-Id: I07fa737d8cf892d27ce08c305dafb0a53fef36cb