Simplify Oryon ifunc resolvers.
Mainly just factoring out the code. The only functional change
here is that the previous code accepted variants 0x0 to 0x5 while
the comment said 0x1 to 0x5.

I resisted the temptation to actually have a table to search on the assumption that it'll be a while before we need such a thing.

Change-Id: I9fdc1e70e49b26ef32794b55ca5e5fd37f1163f9