Revert "Update to v6.8 kernel headers."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7f5b7017-8f2e-4821-9532-35e7df1cc5e8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2cbf12f0ad147dc63fe57edf648ce85d9522ffff